Improve IsCanonical matching.
The RFC leaves is intentionally vague if an HTTP redirect determines
whether or not a file can be canonical for a domain.  Thus, we will
allow that now.  Now, if a "security.txt" file is tested for canonicity,
it will be deemed canonical if a Canonical token matches either the
user-supplied URI or the URI used to fetch the resource.

However, per the RFc, the user should be making the final judgement
call.  (We're just here to help.)