Add Windows PowerShell compatibility.
This module claimed to be compatible with Windows PowerShell, but was
not.  It was relying on the IsSuccessStatusCode property to have a
value, but that is not supported on .NET Framework apparently.

By changing this to an explicit status code numerical check, we have
restored compatibility with Windows PowerShell 5.1 (for anyone still
using that old but popular version).