Modified Signature checks
This adds signature checks via the use of `status-fd`

Notes:
* This will probably obsololte the checking of different output stream
  since status-fd will specify which one to check
* The use of `[...].Matches.Groups[1].Value` shouldn't cause an error.
  (Or at least it didn't in my testing) it would be possible to rewrite
  this to handle strange edge case which could cause a null index error

WARNING:
this may break Signature Checking for keys which aren't already within
the current keystore.

The `correct` way to handle that should be to import said key before
checking the signature.

This could maybe be handled within this module but that would _probably_
fall out of the scope of this PR.