Edit PR #4 to change the definition of "good" signature.
Thank you for your hard work on this pull request. In the context
of this module, I consider a "good" signature to be one that is
valid, regardless of whether or not the key is in GPG's keyring.
This is in contrast to what GPG calls a "good" signature.

I've corrected your code to comply with my different definition,
and I've also enabled it to properly report bad signatures, too.
I will add documentation notes when I can.